Attempt to nicely OOP up, but it's still ehh
